+++ tests/IntegrationTests/WebAppCallsApiCallsGraphUiTests/WebAppCallsApiCallsGraphUiTests.csproj
@@ -12,7 +12,6 @@
     <PackageReference Include="Microsoft.Identity.Lab.Api" Version="$(MicrosoftIdentityLabApiVersion)" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="$(MicrosoftNetTestSdkVersion)" />
     <PackageReference Include="Microsoft.Playwright" Version="$(MicrosoftPlaywrightVersion)" />
-    <PackageReference Include="System.Management" Version="7.0.2" />
     <PackageReference Include="xunit" Version="$(XunitVersion)" />
     <PackageReference Include="xunit.runner.visualstudio" Version="$(XunitRunnerVisualStudioVersion)">
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
