+++ tests/IntegrationTests/WebAppUiTests/TestingWebAppLocally.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using System;
+using System.Collections.Generic;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
 using System.Threading.Tasks;
@@ -16,32 +17,21 @@ public class TestingWebAppLocally
 {
     private const string UrlString = "https://localhost:5001/MicrosoftIdentity/Account/signin";
     private const string DevAppPath = @"DevApps\WebAppCallsMicrosoftGraph";
-    private const string DevAppExecutable = "WebAppCallsMicrosoftGraph.exe";
+    private const string DevAppExecutable = @"\WebAppCallsMicrosoftGraph.exe";
     private string UiTestAssemblyLocation = typeof(TestingWebAppLocally).Assembly.Location;
 
     [Fact]
     public async Task ChallengeUser_MicrosoftIdentityFlow_LocalApp_ValidEmailPasswordCreds_SignInSucceedsTestAsync()
     {
-/*
-        // Uncomment to initialise Playwright, which will download the browsers
-        var exitCode = Microsoft.Playwright.Program.Main(new[] { "install" });
-        if (exitCode != 0)
-        {
-            throw new Exception($"Playwright exited with code {exitCode}");
-        }
-*/
         if (!RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
         { return; }
 
         // Arrange
         Process? p = UiTestHelpers.StartWebAppLocally(UiTestAssemblyLocation, DevAppPath, DevAppExecutable);
 
-        if (p != null)
+        try
         {
-            if (p.HasExited)
-            {
-                Assert.Fail($"Could not run web app locally.");
-            }
+            if (p == null || p.HasExited) { Assert.Fail($"Could not run web app locally."); }
 
             using var playwright = await Playwright.CreateAsync();
             IBrowser browser;
@@ -50,25 +40,22 @@ public class TestingWebAppLocally
             await page.GotoAsync(UrlString);
             LabResponse labResponse = await LabUserHelper.GetDefaultUserAsync().ConfigureAwait(false);
 
-            try
-            {
-                // Act
-                Trace.WriteLine("Starting Playwright automation: web app sign-in & call Graph");
-                string email = labResponse.User.Upn;
-                await UiTestHelpers.FirstLogin_MicrosoftIdentityFlow_ValidEmailPassword(page, email, labResponse.User.GetOrFetchPassword());
+            // Act
+            Trace.WriteLine("Starting Playwright automation: web app sign-in & call Graph");
+            string email = labResponse.User.Upn;
+            await UiTestHelpers.FirstLogin_MicrosoftIdentityFlow_ValidEmailPassword(page, email, labResponse.User.GetOrFetchPassword());
 
-                // Assert
-                await Assertions.Expect(page.GetByText("Welcome")).ToBeVisibleAsync();
-                await Assertions.Expect(page.GetByText(email)).ToBeVisibleAsync();
-            }
-            catch (Exception ex)
-            {
-                Assert.Fail($"the UI automation failed: {ex}");
-            }
-            finally
-            {
-                p.Kill(true);
-            }
+            // Assert
+            await Assertions.Expect(page.GetByText("Welcome")).ToBeVisibleAsync();
+            await Assertions.Expect(page.GetByText(email)).ToBeVisibleAsync();
+        } catch (Exception ex)
+        {
+            Assert.Fail($"the UI automation failed: {ex} output: {ex.Message}");
+        } finally
+        {
+            Queue<Process> processes = new Queue<Process>();
+            processes.Enqueue(p);
+            UiTestHelpers.killProcessTrees(processes);
         }
     }
 }
